dev_environment/requirements.txt
# Minimal requirements for batch-only extraction with GCS support
pandas>=1.5.0
pymysql>=1.0.0
pyarrow>=10.0.0
google-cloud-storage>=2.10.0
google-auth>=2.22.0
google-cloud-bigquery>=3.11.0
apache-airflow>=2.5.0  # For Cloud Composer